Fetch all keys by default
